+++ ci4/app/Entities/AutoUpdate.php
@@ -39,7 +39,7 @@ class AutoUpdate extends Entity {
             ->find();
 
         foreach ($deployments as $deployment) {
-            if ($deployment->workspace->status == \WorkspaceStatusTypes::Inactive) {
+            if ($deployment->isInAPausedOrInactiveWorkspace()) {
                 continue;
             }
             if (preg_match("/{$deployment->auto_update_tag_regex}$/", $tag)) {
@@ -81,8 +81,8 @@ class AutoUpdate extends Entity {
         $deployment = new Deployment();
         $deployment->find($this->deployment_id);
 
-        if ($deployment->workspace->status == \WorkspaceStatusTypes::Inactive) {
-            Data::debug('Skip rollout because workspace is inactive');
+        if ($deployment->isInAPausedOrInactiveWorkspace()) {
+            Data::debug('Skip rollout because the workspace is paused or inactive');
             return;
         }